Copy time metadata from source data
This makes blending also work when doing it mid-timeseries.
time_offset has been changed to seconds and default values is
3600, meaning that if source data has for example analysis hour
08, the blended data analysis time is shifted 1 hour to 09.
This is how the current smartmet nwc works.

For other data producers, like hrnwc, backwards compatibility is
not quaranteed. For hrnwc the following options need to be added
to provide the same output as currently:

--seconds_between_step 900 --time_offset 0